requirements.txt
click
pymysql
pyyaml
sqlalchemy
